third_party/blink/web_tests/wpt_internal/origin_trials/coep_credentialless/driver-without-coep.https-expected.txt

This is a testharness.js-based test.
PASS Install the third party cookies.
PASS environment="document" origin_trial="enabled"
FAIL environment="document" origin_trial="disabled" assert_equals: expected (string) "cookie" but got (undefined) undefined
PASS environment="service_worker" origin_trial="enabled"
FAIL environment="service_worker" origin_trial="disabled" assert_equals: expected (string) "cookie" but got (undefined) undefined
FAIL environment="dedicated_worker" origin_trial="enabled" assert_equals: expected (undefined) undefined but got (string) "cookie"
PASS environment="dedicated_worker" origin_trial="disabled"
PASS environment="shared_worker" origin_trial="enabled"
FAIL environment="shared_worker" origin_trial="disabled" assert_equals: expected (string) "cookie" but got (undefined) undefined
Harness: the test ran to completion.
